% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_vers_PD.R
\name{new_user_inputs}
\alias{new_user_inputs}
\title{Proteome Discoverer analysis}
\usage{
new_user_inputs(
  excel_file,
  groups_number,
  group_names,
  case_number,
  imputation = TRUE,
  global_threshold = TRUE,
  MWtest = "Independent",
  threshold_value = 50,
  normalization = FALSE
)
}
\arguments{
\item{excel_file}{The whole path to the excel .xlsx file, that will be analysed. Attention: Add '/' between the directories.}

\item{groups_number}{The number of the different groups}

\item{group_names}{The names attributed to each different group. Insert in form of a vector. The order of the names should align with the order in the inserted excel file.}

\item{case_number}{The number of samples attributed to each different group. Insert in form of a vector. The order of the number of groups should align with the order in the inserted excel file.}

\item{imputation}{TRUE/FALSE Data imputation using kNN classification or assigning missing values as 0.}

\item{global_threshold}{TRUE/FALSE If TRUE threshold for missing values will be applied to the groups altogether, if FALSE to each group seperately}

\item{MWtest}{Either "Paired" for a Wilcoxon Signed-rank test or "Independent" for a Mann-Whitney U test.}

\item{threshold_value}{The percentage of missing values per protein that will cause its deletion}

\item{normalization}{PPM, TIC etc.}
}
\value{
Excel files with the proteomic values from all samples, processed with normalization and imputation and substraction of samples with high number of missing values. PCA plots for all or for just the significant correlations, and boxplots for the proteins of each sample.
}
\description{
It takes Proteomics Data from samples in different groups, in the format they are created by Proteome Discoverer (PD). It concatenates the Accession IDs, Descriptions and Areas from different PD export files into a master table and performs exploratory data analysis. The function outputs a normalized Parts Per Million protein dataset along with descriptive statistics and results of significance testing. The script also creates exploratory plots such as relative log espression boxplots and PCA plots.
}
\examples{
#' # Example of running the function with paths for two groups.
#Do not add if (interactive()){} condition in your code
if (interactive()){
new_user_inputs(excel_file = "C:/Users/User/Documents/itern/new_version_PD.xlsx",
groups_number = 2,group_names = c("exp_2020","exp_2024"),
 case_number = c(18,8), imputation = FALSE, threshold_value = 100)}

}
