% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_res.R
\name{ProtoAnalysis}
\alias{ProtoAnalysis}
\title{Proteome Discoverer analysis}
\usage{
ProtoAnalysis(
  ...,
  imputation = TRUE,
  global_threshold = TRUE,
  MWtest = "Independent",
  threshold_value = 50,
  bugs = 0,
  normalization = FALSE,
  parametric = FALSE,
  significancy = "pV"
)
}
\arguments{
\item{...}{The specific path to the folder where the samples from each group are located. They are passed as unnamed arguments via "...".  Attention: Add '/' between the directories.}

\item{imputation}{TRUE/FALSE Data imputation using kNN classification or assigning missing values as 0.}

\item{global_threshold}{TRUE/FALSE If TRUE threshold for missing values will be applied to the groups altogether, if FALSE to each group seperately}

\item{MWtest}{Either "Paired" for a Wilcoxon Signed-rank test or "Independent" for a Mann-Whitney U test.}

\item{threshold_value}{The percentage of missing values per protein that will cause its omittion.}

\item{bugs}{Either 0 to treat Proteome Discoverer bugs as Zeros (0) or "average" to convert them into the average of the protein between the samples.}

\item{normalization}{PpM, Quantile, VST}

\item{parametric}{TRUE/FALSE Choose which statistical test will be taken into account when creating the optical statistical analysis (PCA plots, heatmap)}

\item{significancy}{pV or adj.pV Choose if the significant values for the PCA plots and the heatmap will derive from the pValue or the adjusted pValue of the comparison.}
}
\value{
Excel files with the proteomic values from all samples, processed with normalization and imputation and substraction of samples with high number of missing values. PCA plots for all or for just the significant correlations, and boxplots for the proteins of each sample.
}
\description{
It takes Proteomics Data from samples in different groups, in the format they are created by Proteome Discoverer (PD). It concatenates the Accession IDs, Descriptions and Areas from different PD export files into a master table and performs exploratory data analysis. The function outputs a normalized Parts Per Million protein dataset along with descriptive statistics and results of significance testing. The script also creates exploratory plots such as relative log espression boxplots and PCA plots.
}
\examples{

# Paths to example data folders within the package
T1_path <- system.file("extdata", "PDexports(single_file)",
 "T1_BLCA", package = "PACKAGE")
T2_path <- system.file("extdata", "PDexports(single_file)",
"T2_BLCA", package = "PACKAGE")
Ta_path <- system.file("extdata", "PDexports(single_file)",
"Ta_BLCA", package = "PACKAGE")

# Run the function with these paths
ProtoAnalysis(
  T1_path, T2_path, Ta_path,
  normalization = "PPM", MWtest = "Paired",
  imputation = "LOD", global_threshold = TRUE
)


}
