% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAnn.R
\name{dianno}
\alias{dianno}
\title{DIA NN  proteomics data analysis}
\usage{
dianno(
  file,
  group_names,
  samples_per_group,
  imputation = FALSE,
  global_filtering = TRUE,
  sample_relationship = "Independent",
  threshold_value = 50,
  parametric = FALSE,
  significance = "pV",
  description = FALSE
)
}
\arguments{
\item{file}{The whole path to the DIA-NN .tsv files (pg_matrix or unique_genes_matrix), that will be analyzed. Attention: Ensure to use forward slashes (/) for specifying paths.}

\item{group_names}{The names attributed to each different group. Insert in form of a vector. The order of the names should align with the order in the inserted tsv file.}

\item{samples_per_group}{The number of samples attributed to each different group. Insert in form of a vector. The order of the number of groups should align with the order in the inserted tsv file.}

\item{imputation}{Imputation of the Missing Values. By default it is set to FALSE. Options are FALSE for no imputation implemented, "LOD" for assigning the lowest protein intensity identified to each MV and "LOD/2" to apply the half of it. Option "kNN" performs a default kNN imputation and "missRanger" a missRanger one. This 2 options are combined with a boxplot that visualizes the distribution of the log2 intensities of the imputed data compared to the initial ones.}

\item{global_filtering}{TRUE/FALSE If TRUE threshold for missing values will be applied to the groups altogether, if FALSE to each group separately}

\item{sample_relationship}{Either "Independent" when the samples come from different populations or "Paired" when they come from the same. By default, it is set to "Independent". If "Paired" is selected the samples_per_group must be equal to each other}

\item{threshold_value}{The percentage of missing values per protein that will cause its omission. By default it is set to 50. (50 percent)}

\item{parametric}{TRUE/FALSE Choose which statistical test will be taken into account when creating visualization of the features. By default it is set to FALSE (non Parametric)}

\item{significance}{pV or adj.pV Choose if the significant values for the PCA plots and the heatmap will derive from the unadjusted pValue or the adjusted pValue (Benjamini-Hochberg) of the comparison. By default it is set "pV" (pValue)}

\item{description}{If TRUE protein information about the first protein inside the protein group like protein name, SV, OS, organism. These option is applicable only when a pg.matrix (protein groups) tsv is provided and requires internet access. By default it is set to FALSE (No description fetching)}
}
\value{
Excel files with the proteomic values that are optionally processed, via imputation and the filtering of proteins with a selected percentage of missing values. The result of the processing is visualized with an Protein Rank Abundance plot. PCA plots for all groups and for just their significant correlations are created. Furthermore violin and boxplots for the proteins of each sample is created and a heatmap for the significant proteins.
}
\description{
Takes as input Proteomics Data (output of DIA-NN) in the format of a tsv files pg_matrix.tsv and unique_gene_matrix.tsv . Performs exploratory data analysis, while providing different options for data manipulation (description fetching from UniProt, filtering based on the missing values and imputation) It then proceeds to perform statistical analysis, while creating exploratory plots such as relative log expression boxplots and violin plots, heatmaps and PCA plots.
}
\examples{
#Example of running the function with paths for two groups.
\donttest{
report.pg_matrix <- system.file("extdata", "report.pg_matrix.tsv", package = "ProtE")
 dianno(file = report.pg_matrix,
 group_names= c("Healthy","Patients"), samples_per_group= c(5,5),
   global_filtering = TRUE, sample_relationship = "Independent",
 threshold_value = 50, description = FALSE, imputation = FALSE)}

}
