% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAnn.R
\name{dianno}
\alias{dianno}
\title{DIA NN  proteomics data analysis}
\usage{
dianno(
  excel_file,
  group_names,
  case_number,
  imputation = FALSE,
  global_threshold = TRUE,
  sample_relationship = "Independent",
  threshold_value = 50,
  parametric = FALSE,
  significancy = "pV",
  description = FALSE
)
}
\arguments{
\item{excel_file}{The whole path to the excel .xlsx file, that will be analysed. Attention: Add '/' between the directories.}

\item{group_names}{The names attributed to each different group. Insert in form of a vector. The order of the names should align with the order in the inserted excel file.}

\item{case_number}{The number of samples attributed to each different group. Insert in form of a vector. The order of the number of groups should align with the order in the inserted excel file.}

\item{imputation}{TRUE/FALSE Data imputation using kNN classification or assigning missing values as 0.}

\item{global_threshold}{TRUE/FALSE If TRUE threshold for missing values will be applied to the groups altogether, if FALSE to each group seperately}

\item{sample_relationship}{Either "Paired" for a Wilcoxon Signed-rank test or "Independent" for a Mann-Whitney U test.}

\item{threshold_value}{The percentage of missing values per protein that will cause its deletion}

\item{parametric}{TRUE/FALSE Choose which statistical test will be taken into account when creating the optical statistical analysis (PCA plots, heatmap)}

\item{significancy}{pV or adj.pV Choose if the significant values for the PCA plots and the heatmap will derive from the pValue or the adjusted pValue of the comparison.}

\item{description}{If TRUE protein information about the first protein inside the protein group like protein name, SV, OS, organism}
}
\value{
Excel files with the proteomic values that are optionally processed, via imputation and the filtering of proteins with a selected percentage of missing values. The result of the processing is optimized with an Protein Rank Abundance plot. PCA plots for all groups and for just their significant correlations are created. Furthermore violin and boxplots for the proteins of each sample is created and a heatmap for the significant proteins.
}
\description{
It takes as input the Proteomics Data (output of DIA-NN) in the format of an excel file that contains the information for each sample. Then it performs exploratory data analysis. The options for the data manipulation include filtering based on the missing values per protein, fetching their description information from UniProt and their imputation, as well as a quality check with the percentage of MVs across every protein is provided. It then proceeds to perform statistical analysis using the Mann Whitney and the limma t-test for pairwise comparisons and  also Kruskal-Wallis and limma-ANOVA statistical tests,when there are more than 2 groups. The function also creates exploratory plots such as relative log espression boxplots and violin plots, heatmaps of the significant differentially expressed proteins and PCA plots.
}
\examples{
report.pg_matrix <- system.file("extdata/DIA-NNorFragPipeExports.pg.matrix",
 "jittered.pg_matrix.xlsx", package = "PACKAGE")
 dianno(report.pg_matrix,
 group_names= c("Healthy","Patients"), case_number= c(9,9),
   global_threshold = TRUE, sample_relationship = "Independent",
 threshold_value = 50, description = FALSE, imputation = FALSE)
}
