% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_res.R
\name{pd_multi}
\alias{pd_multi}
\title{Proteome Discoverer (PD) multiple-files' proteomic analysis}
\usage{
pd_multi(
  ...,
  imputation = TRUE,
  global_filtering = TRUE,
  sample_relationship = "Independent",
  threshold_value = 50,
  bugs = 0,
  normalization = FALSE,
  parametric = FALSE,
  significancy = "pV"
)
}
\arguments{
\item{...}{The specific path to the folder where the samples from each group are located. They are passed as unnamed arguments via "...".  Attention: Ensure paths use '/' as a directory separator.}

\item{imputation}{Imputation of the Missing Values. Options are "LOD" for assigning the lowest protein intensity identified to each MV and "LOD/2" to apply the half of it. Option "kNN" performs a default kNN imputation and "missRanger" a missRanger one. This 2 options are combined with a boxplot that optimizes the distribution of the log2 itensities of the imputed data compared to the initial ones.}

\item{global_filtering}{TRUE/FALSE If TRUE the threshold for missing values filtering will be applied to the groups altogether, if FALSE it will be applied to each group seperately.}

\item{sample_relationship}{Either "Paired" for a Wilcoxon Signed-rank test or "Independent" for a Mann-Whitney U test.}

\item{threshold_value}{The percentage of missing values per protein that will cause its omittion.}

\item{bugs}{Either 0 to treat Proteome Discoverer bugs as Zeros (0) or "average" to convert them into the average of the protein between the samples.}

\item{normalization}{The specific method for normalizing the data. Options are "log2" for a simple log2 transformation, "Quantile" for a quantiles based normalization, "median" for a median one, "TIC" for Total Ion Current normalization, "VSN" for Variance Stabilizing Normalization and "PPM" for Parts per Million transformation of the data.}

\item{parametric}{TRUE/FALSE Choose which statistical test will be taken into account when creating the optical statistical analysis (PCA plots, heatmap). If TRUE the limma t.test (for 2 groups) or ANOVA (for > 2 groups) significancies will be considered and if FALSE the Mann Whitney (for 2 groups) and Kruskal Wallis (>2 groups)}

\item{significancy}{"pV" or "adj.pV" Choose if the significant values for the PCA plots and the heatmap will derive from the pValue or the adjusted pValue of the comparison.}
}
\value{
Excel files with the proteomic values that are optionally processed, via normalization, imputation and filtering of proteins with a selected percentage of missing values. The result of the processing is optimized with an Protein Rank Abundance plot. PCA plots for all groups and for just their significant correlations are created. Furthermore violin and boxplots for the proteins of each sample is created and a heatmap for the significant proteins.
}
\description{
It takes as input the  Proteomics Data (output of PD) in the format of a single file per sample. After concatenating the Accession_ID, Description and Area or Abundance from the PD export files of each sample, into a master table, it performs exploratory data analysis. The options for the data manipulation include different methods of normalization, and filtering based on the missing values per protein. Additionally, imputation of  the missing values can be performed, and a quality check with their percentage across every protein is provided. It then proceeds to perform statistical analysis using the Mann Whitney and the limma t-test for pairwise comparisons and  also Kruskal-Wallis and limma-ANOVA statistical tests,when there are more than 2 groups, while the pValues from the Levene and Bartlett statistical tests are also shown. The function also creates exploratory plots such as relative log espression boxplots and violin plots, heatmaps of the significant differentially expressed proteins and PCA plots.
}
\examples{

# Paths to example data folders within the package
T1_path <- system.file("extdata", "PDexports(multiple_files)",
 "T1_BLCA", package = "ProtE")
T2_path <- system.file("extdata", "PDexports(multiple_files)",
"T2_BLCA", package = "ProtE")
Ta_path <- system.file("extdata", "PDexports(multiple_files)",
"Ta_BLCA", package = "ProtE")

# Run the function with these paths
pd_multi(
  T1_path, T2_path, Ta_path,
  normalization = "PPM", sample_relationship = "Paired",
  imputation = "LOD", global_filtering = TRUE
)


}
