% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_vers_PD.R
\name{pd_single}
\alias{pd_single}
\title{Proteome Discoverer proteomic data analysis}
\usage{
pd_single(
  excel_file,
  group_names,
  samples_per_group,
  imputation = FALSE,
  global_filtering = TRUE,
  sample_relationship = "Independent",
  threshold_value = 50,
  normalization = FALSE,
  parametric = FALSE,
  significancy = "pV"
)
}
\arguments{
\item{excel_file}{The whole path to the excel *.xlsx file, that will be analysed. Attention: Ensure to use forward slashes (/) for specifying paths.}

\item{group_names}{The names attributed to each different group. Insert in form of a vector. The order of the names should align with the order in the inserted excel file.}

\item{samples_per_group}{The number of samples attributed to each different group. Insert in form of a vector. The order of the number of groups should align with the order in the inserted excel file.}

\item{imputation}{Imputation of the Missing Values. By default it is set to FALSE. Options are FALSE for no imputation implemented, "LOD" for assigning the lowest protein intensity identified to each MV and "LOD/2" to apply the half of it. Option "kNN" performs a default kNN imputation and "missRanger" a missRanger one. This 2 options are combined with a boxplot that visualizes the distribution of the log2 itensities of the imputed data compared to the initial ones.}

\item{global_filtering}{TRUE/FALSE If TRUE the threshold for missing values filtering will be applied to the groups altogether, if FALSE it will be applied to each group separately.}

\item{sample_relationship}{Either "Independent" when the samples come from different populations or "Paired" when they come from the same. By default, it is set to "Independent". If "Paired" is selected the samples_per_group must be equal to each other}

\item{threshold_value}{The percentage of missing values per protein that will cause its omission. By default it is set to 50. (50%)}

\item{normalization}{The specific method for normalizing the data.By default it is set to FALSE. Options are FALSE for no normalization of the data, "log2" for a simple log2 transformation, "Quantile" for a quantiles based normalization, "median" for a median one, "TIC" for Total Ion Current normalization, "VSN" for Variance Stabilizing Normalization and "PPM" for Parts per Million transformation of the data.}

\item{parametric}{TRUE/FALSE Choose which statistical test will be taken into account when creating the optical statistical analysis (PCA plots, heatmap). By default it is set to FALSE (non Parametric)}

\item{significancy}{pV or adj.pV Choose if the significant values for the PCA plots and the heatmap will derive from the unadjusted pValue or the adjusted pValue (Benjamini-Hochberg) of the comparison. By default it is set "pV" (pValue)}
}
\value{
Excel files with the proteomic values that are optionally processed, via normalization , imputation and  filtering of proteins with a selected percentage of missing values. The result of the processing is visualized with an Protein Rank Abundance plot. PCA plots for all groups and for just their significant correlations are created. Furthermore violin and boxplots for the proteins of each sample is created and a heatmap for the significant proteins.
}
\description{
It takes as input the Proteomics Data (output of PD) in the format of an excel file that contains the information for each sample. Then it performs exploratory data analysis. The options for the data manipulation include different methods of normalization, and filtering based on the missing values per protein. Additionally, imputation of  the missing values can be performed, and a quality check with their percentage across every protein is provided. It then proceeds to perform statistical analysis using the Mann Whitney and the limma t-test for pairwise comparisons and  also Kruskal-Wallis and limma-ANOVA statistical tests,when there are more than 2 groups, while the pValues from the Levene and Bartlett statistical tests are also shown. The function also creates exploratory plots such as relative log espression boxplots and violin plots, heatmaps of the significant differentially expressed proteins and PCA plots.
}
\examples{
#' # Example of running the function with paths for two groups.
if (interactive()){
pd_single(excel_file = "C:/Users/User/Documents/new_version_PD.xlsx",
groups_number = 2,group_names = c("exp_2020","exp_2024"),
 samples_per_group = c(18,8), imputation = FALSE, threshold_value = 100)}

}
