% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_res.R
\name{user_inputs}
\alias{user_inputs}
\title{Get data from 2 different sample groups from protein discoverer and create a dataspace wwith all the protein IDs found.}
\usage{
user_inputs(group1, group2, imputation = TRUE, global_threshold = TRUE, MWtest)
}
\arguments{
\item{group1}{path to the folder where the samples from group 1 are}

\item{group2}{path to the folder where the samples from group 2 are}

\item{imputation}{if imputation will happen}

\item{global_threshold}{whether trheshold for missing values will apply to all groups together or to each group seperately}

\item{MWtest}{mann whitney test: Matched for a paired one, Indepedent for an unpaired}
}
\value{
excel file
}
\description{
Creates an a excel file
}
\examples{
user_inputs(group1, group2)
}
