% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_res.R
\name{user_inputs}
\alias{user_inputs}
\title{Proteome Discoverer analysis}
\usage{
user_inputs(..., imputation = TRUE, global_threshold = TRUE, MWtest = "Paired")
}
\arguments{
\item{...}{The specific path to the folder where the samples from each group are located. They are passed as unnamed arguments via "...".  Attention: Add '\\' between the directories and not '/'.}

\item{imputation}{TRUE/FALSE Data imputation using kNN classification or assigning missing values as 0.}

\item{global_threshold}{TRUE/FALSE If TRUE threshold for missing values will be applied to the groups altogether, if FALSE to each group seperately}

\item{MWtest}{Either "Paired" for a Wilcoxon Signed-rank test or "Indepedent" for a Mann-Whitney U test.}
}
\value{
Excel files with the proteomic values from all samples, processed with normalization and imputation and substraction of samples with high number of missing values. PCA plots for all or for just the significant correlations, and boxplots for the proteins of each sample.
}
\description{
It takes Proteomics Data from samples in different groups, in the format they are created by protein discoverer. It processes and analyze these Data in order to create PCA plots and Boxplots.
}
\examples{
#' # Example of running the function with paths for two groups.
#Do not add if (interactive()){} condition in your code
if (interactive()){
user_inputs(
  "C:/Users/User/Documents/T0_samples",
  "C:/Users/User/Documents/T0_samples",
  MWtest = "Paired",
  imputation = TRUE,
  global_threshold = TRUE
)}

}
