% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_res.R
\name{user_inputs}
\alias{user_inputs}
\title{Get data from 2 different sample groups from protein discoverer and create a dataspace wwith all the protein IDs found.}
\usage{
user_inputs(group1, group2, imputation = TRUE, global_threshold = TRUE)
}
\arguments{
\item{group1}{path to the folder where the samples from group 1 are}

\item{group2}{path to the folder where the samples from group 2 are}
}
\value{
excel file
}
\description{
Creates an a excel file
}
\examples{
user_inputs(group1, group2)
}
